+++ ktlint/src/test/kotlin/com/pinterest/ktlint/SimpleCLITest.kt
@@ -173,6 +173,21 @@ class SimpleCLITest {
     }
 
     @Test
+    fun `Issue 1793 - Given some code with an error and no patterns read in from stdin then return nothing `(
+        @TempDir
+        tempDir: Path,
+    ) {
+        CommandLineTestRunner(tempDir)
+            .run(
+                "too-many-empty-lines",
+                listOf("--patterns-from-stdin"),
+                stdin = ByteArrayInputStream(ByteArray(0)),
+            ) {
+                assertNormalExitCode()
+            }
+    }
+
+    @Test
     fun `Issue 1608 - --relative and --reporter=sarif should play well together`(
         @TempDir
         tempDir: Path,